+++ packages/brick_offline_first_with_rest_build/test/offline_first_generator/test_rest_config_response_keys.dart
@@ -48,14 +48,14 @@ class RestConfigResponseKeysAdapter
   final Map<String, Map<String, dynamic>> fieldsToSqliteColumns = {
     'primaryKey': SqliteColumnDefinition(
       association: false,
+      columnName: '_brick_id',
       iterable: false,
-      name: '_brick_id',
       type: int,
     ),
     'someField': SqliteColumnDefinition(
       association: false,
+      columnName: 'some_field',
       iterable: false,
-      name: 'some_field',
       type: int,
     ),
   };
