+++ packages/brick_sqlite/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Change `SqliteAdpater#fieldsToSqliteColumns` type from `Map<String, Map<String, dynamic>>` to `Map<String, SqliteColumnDefinition>`. Using such a dynamic type option will lead to inconsistency when accessing the values.
+
 ## 0.1.5+1
 
 * Fix: recreate SQLite DB after reset instead of attempting to open a closed DB
