+++ packages/brick_sqlite_generators/test/sqlite_model_serdes_generator/test_field_with_type_argument.dart
@@ -29,16 +29,16 @@ class FieldWithTypeArgumentAdapter
   FieldWithTypeArgumentAdapter();
 
   final Map<String, Map<String, dynamic>> fieldsToSqliteColumns = {
-        'primaryKey': SqliteColumnDefinition(
+    'primaryKey': SqliteColumnDefinition(
       association: false,
+      columnName: '_brick_id',
       iterable: false,
-      name: '_brick_id',
       type: int,
     ),
     'someField': SqliteColumnDefinition(
       association: false,
+      columnName: 'some_field',
       iterable: false,
-      name: 'some_field',
       type: Map,
     ),
   };
