+++ packages/brick_sqlite_generators/test/sqlite_model_serdes_generator/test_sqlite_unique.dart
@@ -25,14 +25,14 @@ class SqliteUniqueAdapter extends SqliteAdapter<SqliteUnique> {
   final Map<String, Map<String, dynamic>> fieldsToSqliteColumns = {
     'primaryKey': SqliteColumnDefinition(
       association: false,
+      columnName: '_brick_id',
       iterable: false,
-      name: '_brick_id',
       type: int,
     ),
     'someField': SqliteColumnDefinition(
       association: false,
+      columnName: 'some_field',
       iterable: false,
-      name: 'some_field',
       type: int,
     ),
   };
